requirements.txt
loguru~=0.6.0
aiohttp~=3.8.1
toml~=0.10.2
aiomysql~=0.1.1
pypika~=0.48.9